// src/app/openapi/confirmSignUp.ts
/**
 * Apart-Deal API Contracts
 * Apart-Deal API Contracts
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ConfirmSignUp { 
    code: string;
    token: string;
}